[BugFix] Fix CI failures in test_setup and REDQ benchmark (#3559)

+++ test/test_loggers.py
@@ -107,7 +107,10 @@ class TestTensorboard:
         # C - number of image channels (e.g. 3 for RGB), H, W - image dimensions.
         # the first 64 frames are black and the next 64 are white
         video = torch.cat(
-            (torch.zeros(64, 1, 32, 32), torch.full((64, 1, 32, 32), 255))
+            (
+                torch.zeros(64, 1, 32, 32, dtype=torch.uint8),
+                torch.full((64, 1, 32, 32), 255, dtype=torch.uint8),
+            )
         )
         video = video[None, :]
         for i in range(3):
@@ -303,7 +306,10 @@ class TestWandbLogger:
         # C - number of image channels (e.g. 3 for RGB), H, W - image dimensions.
         # the first 64 frames are black and the next 64 are white
         video = torch.cat(
-            (torch.zeros(128, 1, 32, 32), torch.full((128, 1, 32, 32), 255))
+            (
+                torch.zeros(128, 1, 32, 32, dtype=torch.uint8),
+                torch.full((128, 1, 32, 32), 255, dtype=torch.uint8),
+            )
         )
         video = video[None, :]
         wandb_logger.log_video(
@@ -390,7 +396,10 @@ class TestMLFlowLogger:
 
         logger, client = mlflow_fixture
         videos = torch.cat(
-            (torch.full((3, 64, 3, 32, 32), 255), torch.zeros(3, 64, 3, 32, 32)),
+            (
+                torch.full((3, 64, 3, 32, 32), 255, dtype=torch.uint8),
+                torch.zeros(3, 64, 3, 32, 32, dtype=torch.uint8),
+            ),
             dim=1,
         )
         fps = 6
@@ -506,7 +515,10 @@ class TestTrackioLogger:
         # C - number of image channels (e.g. 3 for RGB), H, W - image dimensions.
         # the first 64 frames are black and the next 64 are white
         video = torch.cat(
-            (torch.zeros(128, 3, 32, 32), torch.full((128, 3, 32, 32), 255))
+            (
+                torch.zeros(128, 3, 32, 32, dtype=torch.uint8),
+                torch.full((128, 3, 32, 32), 255, dtype=torch.uint8),
+            )
         )
         video = video[None, :]
         trackio_logger.log_video(
